fix(task): close lifecycle publication findings
