+++ packages/kernel/src/composition/index.ts
@@ -4,6 +4,6 @@ export { readContentAddressedBlob, readContentAddressedTextBlob, writeContentAdd
 export { makeMarkdownArtifactStore } from "../store/markdown-artifact-store.ts";
 export { makeJournaledWriteCoordinator } from "../store/write-journal-coordinator.ts";
 export { makeLocalLockRegistry } from "../store/local-lock-registry.ts";
-export { makeTaskLeaseStore, TaskLeaseConflictError } from "../local/task-lease-store.ts";
+export { makeTaskLeaseStore } from "../local/task-lease-store.ts";
 export { makeTaskProjection } from "../projection/task-projection.ts";
 export { makeTaskEventStore, TaskEventStoreError } from "../store/task-event-store.ts";
