+++ packages/kernel/src/index.ts
@@ -1,9 +1,7 @@
 export * from "./domain/index.ts";
 export {
   applyTransition,
-  assertAntiEntropyGraph,
-  freezeWritePlan,
-  validateTransition
+  freezeWritePlan
 } from "./domain/task-lifecycle.contract.ts";
 export type {
   CompleteTaskProof,
@@ -14,8 +12,7 @@ export type {
   TaskLifecycleSnapshot
 } from "./domain/task-lifecycle.contract.ts";
 export { canonicalizeContractValue } from "./domain/task.ts";
-export type { FrozenWritePlan, TaskV1 } from "./domain/task.ts";
-export { TASK_LEASE_BROKER_CONTRACT } from "./domain/execution.ts";
+export type { FrozenWritePlan, TaskV1, WriteTarget } from "./domain/task.ts";
 export type { LeaseV1 } from "./domain/execution.ts";
 export * from "./docmap/index.ts";
 export * from "./docmap/docmap-unique.ts";
@@ -79,7 +76,6 @@ export {
   makeMarkdownArtifactStore,
   makeTaskEventStore,
   makeTaskLeaseStore,
-  TaskLeaseConflictError,
   makeTaskProjection,
   readContentAddressedTextBlob,
   writeContentAddressedBlob
