+++ packages/kernel/test/contracts/public-surface.test.ts
@@ -52,10 +52,8 @@ const publicRuntimeSurface = [
   "SnapshotStatusSchema",
   "SqliteTaskRowSchema",
   "SubtaskPlanSchema",
-  "TASK_LEASE_BROKER_CONTRACT",
   "TASK_LIFECYCLE_COMMAND_CATALOG",
   "TaskFrontmatterSchema",
-  "TaskLeaseConflictError",
   "TaskSnapshotSchema",
   "TemplateCatalogSchema",
   "TemplateLibrary",
@@ -66,7 +64,6 @@ const publicRuntimeSurface = [
   "WriteCoordinator",
   "WriteJournalOpSchema",
   "applyTransition",
-  "assertAntiEntropyGraph",
   "assertNoPortablePathCollisions",
   "assertUniqueDocmapIds",
   "auditTaskProvenance",
@@ -210,7 +207,6 @@ const publicRuntimeSurface = [
   "validateRelationRecordsForHost",
   "validateTaskIdSyntax",
   "validateTemplateCatalog",
-  "validateTransition",
   "validateVerticalDefinition",
   "warning",
   "writeContentAddressedBlob",
