+++ packages/kernel/test/store/task-lifecycle-runtime.ts
@@ -1,6 +1,7 @@
 export {
   makeTaskEventStore,
   makeTaskLeaseStore,
-  makeTaskProjection,
-  TaskLeaseConflictError
+  makeTaskProjection
 } from "../../src/composition/index.ts";
+export { TASK_LEASE_BROKER_CONTRACT } from "../../src/domain/execution.ts";
+export { TaskLeaseConflictError } from "../../src/local/task-lease-store.ts";
