+++ tools/gate-allowlists/check-bypass-write-boundary.json
@@ -2,6 +2,103 @@
   "schema": "harness-anything/gate-allowlist/v1",
   "gateId": "check-bypass-write-boundary",
   "entries": {
+    "rebuildable-projection": [
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#DatabaseSync@1",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle SQLite is a disposable projection cache rebuilt exclusively from the authoritative task event stream."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.exec@1",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection rebuild drops only disposable derived tables."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.exec@2",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection configures its disposable SQLite cache."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.exec@3",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection creates disposable derived tables."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.exec@4",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection begins an atomic cache-only transaction."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.exec@5",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection commits an atomic cache-only transaction."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.exec@6",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection rolls back an atomic cache-only transaction."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@1",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection materializes a task event into the disposable cache."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@2",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection materializes an execution into the disposable cache."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@3",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection materializes a review into the disposable cache."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@4",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection materializes a graph edge into the disposable cache."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@5",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection reads its derived watermark."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@6",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection reads a derived task snapshot."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@7",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection compares derived event rows with authoritative events."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@8",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection reads latest disposable snapshots for validation."
+      },
+      {
+        "value": "packages/kernel/src/projection/rebuildable-task-projection.ts#sqlite.prepare@9",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "P4 task lifecycle projection reads disposable entity rows for validation."
+      },
+      {
+        "value": "packages/kernel/src/projection/sqlite-projection-store.ts#mkdirSync@1",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "The entity SQLite projection is a disposable cache rebuilt from canonical task and decision documents."
+      },
+      {
+        "value": "packages/kernel/src/projection/sqlite-projection-store.ts#rmSync@1",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "The entity SQLite projection removes only a disposable temporary cache before rebuilding it."
+      },
+      {
+        "value": "packages/kernel/src/projection/sqlite-projection-store.ts#renameSync@1",
+        "ref": "task_01KZQ944F7RH6E0153Q3Q86Y63",
+        "reason": "The entity SQLite projection atomically publishes a fully rebuildable cache file."
+      }
+    ],
     "coordinatedCore": [
       {
         "value": "packages/kernel/src/store/markdown-artifact-store.ts#mkdirSync@1",
