+++ tools/gates/line-budgets.json
@@ -5,7 +5,7 @@
     "task-lifecycle": 3068,
     "doc-sync": 679,
     "preset": 6926,
-    "cli": 16778,
+    "cli": 16777,
     "gui": 15309,
     "daemon": 2333,
     "authority-write-path": 0,
